More fixes for v11, definitely fix the random freezes on Ideapad 530s (issue #1).
The random freezes were due to some kernel scheduling/locking oddity I couldn't entirely figure out.

But by getting rid of the synchronize_irq call (which was the one stalling, despite the isr having already returned) and simplifying the task flow by removing the delayed_work field (so the thread count is reduced from 3 to 2), the freezes and long spinning in [irq/NN-MSFT0001] appear to be completely gone.